+++ .github/workflows/prep-e2e.yml
@@ -92,8 +92,8 @@ jobs:
             "test-e2e-chrome-flask-webpack": 5,
             "test-e2e-firefox-webpack": 20,
             "test-e2e-firefox-flask-webpack": 5,
-            "test-e2e-chrome-playwright": 5,
-            "test-e2e-firefox-playwright": 5
+            "test-e2e-chrome-playwright": 3,
+            "test-e2e-firefox-playwright": 3
           }')
 
           # Expand counts into the job matrix shape expected by downstream workflows
