+++ test/e2e/playwright/shared/chrome-extension-harness.ts
@@ -82,16 +82,7 @@ export async function launchMetaMaskChromeExtension(
     '--disable-features=Translate',
     '--disable-component-update',
     '--disable-dev-shm-usage',
-    // CRITICAL: trust mockttp's self-signed MITM certificate browser-wide.
-    // Playwright's `ignoreHTTPSErrors: true` below only applies to targets
-    // Playwright attaches to (pages/frames) — it does NOT cover the MV3
-    // extension service worker, so without this flag every HTTPS fetch from
-    // the background (remote feature flags, accounts/price/token APIs, ...)
-    // fails its TLS handshake against the proxy and the extension silently
-    // runs without backend data (e.g. remote feature flags never load and
-    // the UI falls back to pre-redesign variants). Selenium's
-    // `acceptInsecureCerts` capability is browser-wide, which is why the
-    // Selenium path doesn't need this.
+    // Required for the proxy to work as Playwright's per-page HTTPS error ignoring (ignoreHTTPSErrors) doesn't cover MV3 extension service workers
     '--ignore-certificate-errors',
     ...(options.extraArgs ?? []),
   ];
